python/docintel/requirements.txt
azure-ai-formrecognizer
azure-identity
authlib
python-dotenv